SONARKT-794 Split ruling test into per-corpus classes for parallel forks
Gradle's maxParallelForks forks per test class, not per test method, so
the single KotlinRulingTest class made parallelism impossible. Splitting
it into an AbstractKotlinRulingTest base plus one concrete subclass per
corpus lets each corpus run in its own forked JVM/classloader, avoiding
the DefaultProfiler thread-safety issue that made JUnit-level concurrency
within a single fork unsafe. RulingRules is folded into the base class
since it was only used there.
