+++ its/ruling/build.gradle.kts
@@ -33,11 +33,13 @@ tasks.integrationTest {
     systemProperty("gradle.main.compile.classpath", sourceSets.main.get().compileClasspath.asPath)
     // Each SIT run leaks engine class loaders; a fresh JVM per class keeps heap bounded.
     setForkEvery(1)
-    // Single test class, so Gradle-level forking wouldn't help anyway; see junit-platform.properties
-    // for why JUnit-level parallelism isn't an option either (shared, non-thread-safe engine state).
-    maxParallelForks = 1
+    // One test class per corpus (see AbstractKotlinRulingTest), each forked into its own JVM/classloader, so
+    // corpora now run in parallel across forks. JUnit-level parallelism within a single fork is still not an
+    // option; see junit-platform.properties for why (shared, non-thread-safe engine state).
+    maxParallelForks = 2
     // The engine now runs inside this JVM (the orchestrator used to fork a scanner process with its own -Xmx),
-    // and the ruling corpus needs considerably more than the Gradle default.
+    // and the ruling corpus needs considerably more than the Gradle default. Kept per-fork rather than raised
+    // further, since maxParallelForks concurrent forks each get their own heap of this size.
     maxHeapSize = "4g"
 }
